+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Structs and functions for listing accounts in the wallet:
+  - `zcashlc_list_accounts`
+  - `zcashlc_free_accounts`
+  - `FfiAccounts`
+  - `FfiAccount`
+
 ## 0.6.0 - 2024-03-07
 
 ### Added